- Move the serializer of aggregated BLS to the same file with the rest of the code - Use SerializationHelper to serialize a fixed length array
