+++ fastcrypto/src/bls12381/min_pk/mod.rs
@@ -4,7 +4,6 @@
 //! Module minimizing the size of public keys. See also [min_sig].
 
 use super::*;
-use crate::serde_helpers::min_pk::BlsSignature;
 use blst::min_pk as blst;
 /// Hash-to-curve domain separation tag.
 pub const DST_G2: &[u8] = b"BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_NUL_";
